+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v5.0.0] - Unreleased
+
+## Changed
+ - expected log file name was changed to 'source/oxideshop.log'
+ - expected log file format was changed
+
 ## [v4.0.0] - 2018-04-30
 
 ### Deprecated